# apps/api/openapi/minerva.json
{
  "openapi": "3.1.1",
  "paths": {
...
    "/minerva/reviews/prompts": {
      "get": {
        "description": "Every prompt the caller has, archived ones included: daily before weekly, Reflect before Plan, each in its order. The first time a user's prompts are read they are given the starter prompts, once.",
        "operationId": "ListReviewPrompts",
        "parameters": [
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "description": "daily or weekly, for only that kind's prompts",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The caller's prompts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListReviewPromptsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Lists the signed-in user's review prompts",
        "tags": [
          "Review Prompts"
        ]
      },
      "post": {
        "description": "Adds a prompt at the end of its kind and section; reviews of that kind ask it from then on.",
        "operationId": "CreateReviewPrompt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Input for the CreateReviewPrompt operation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewPromptRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The prompt was created.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "The location of the created prompt"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateReviewPromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Another prompt was added to the section at the same time."
          }
        },
        "summary": "Creates a review prompt for the signed-in user",
        "tags": [
          "Review Prompts"
        ]
      }
    },
    "/minerva/reviews/prompts/order": {
      "put": {
        "description": "Puts the caller's prompts of one kind and section in the order given, which must name every one of them, archived ones included, exactly once.",
        "operationId": "ReorderReviewPrompts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Input for the ReorderReviewPrompts operation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderReviewPromptsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "All of the caller's prompts, in their new order.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReorderReviewPromptsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Reorders the signed-in user's review prompts in one section",
        "tags": [
          "Review Prompts"
        ]
      }
    },
    "/minerva/reviews/prompt/{promptId}": {
      "get": {
        "description": "Returns a single prompt of the caller's. Another user's prompt is not found.",
        "operationId": "DescribeReviewPrompt",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "The ID of the prompt",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The prompt was found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescribeReviewPromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Describes one of the signed-in user's review prompts",
        "tags": [
          "Review Prompts"
        ]
      },
      "put": {
        "description": "Rewords a prompt, changes or removes its placeholder, or archives it or brings it back. An archived prompt is no longer asked; reviews that answered it keep it.",
        "operationId": "UpdateReviewPrompt",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "The ID of the prompt to change",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Input for the UpdateReviewPrompt operation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewPromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The prompt with the changes applied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateReviewPromptResponse"
                }
              }
            }
          },
          "304": {
            "description": "The request named nothing to change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Changes one of the signed-in user's review prompts",
        "tags": [
          "Review Prompts"
        ]
      },
      "delete": {
        "description": "Removes a prompt that has never been answered. A prompt with answers is refused; archive it instead, so the reviews that answered it keep it.",
        "operationId": "DeleteReviewPrompt",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "The ID of the prompt to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The prompt was deleted."
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The prompt has answers."
          }
        },
        "summary": "Deletes one of the signed-in user's review prompts",
        "tags": [
          "Review Prompts"
        ]
      }
    },
...
    "/minerva/review/{reviewId}/answer/{promptId}": {
      "put": {
        "description": "Writes or rewrites the answer to a prompt of the caller's, of the review's kind. An answer saved again keeps its created time; one saved after the review was completed is marked as edited later. A body that is empty or only whitespace removes the answer.",
        "operationId": "UpdateReviewAnswer",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "description": "The ID of the review",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "The ID of the prompt answered",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Input for the UpdateReviewAnswer operation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The answer as saved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateReviewAnswerResponse"
                }
              }
            }
          },
          "204": {
            "description": "The body was empty, so there is no answer now."
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Saves the answer to one of a review's prompts",
        "tags": [
          "Reviews"
        ]
      }
    }
  },
  "info": {
    "title": "Minerva API",
    "description": "API documentation for Minerva",
    "version": "0.0.0",
    "contact": {
      "name": "Neil Fritz",
      "url": "https://ncfritz.net",
      "email": "ncfritz@ncfritz.net"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "number"
          }
        },
        "required": [
          "message",
          "statusCode"
        ]
      },
...
      "EmptyResponse": {
        "type": "object",
        "properties": {}
      },
...
      "ReviewKind": {
        "type": "string",
        "description": "Whether a review is of a day or of an ISO week",
        "enum": [
          "daily",
          "weekly"
        ]
      },
      "ReviewPromptSection": {
        "type": "string",
        "description": "The guided step a review prompt is asked in",
        "enum": [
          "reflect",
          "plan"
        ]
      },
      "ReviewPrompt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the prompt"
          },
          "kind": {
            "description": "The kind of review that asks the prompt",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewKind"
              }
            ]
          },
          "section": {
            "description": "The step the prompt is asked in",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPromptSection"
              }
            ]
          },
          "label": {
            "type": "string",
            "description": "The question, 1 to 120 characters"
          },
          "placeholder": {
            "type": "string",
            "description": "A hint shown in the empty answer box, at most 200 characters"
          },
          "position": {
            "type": "number",
            "description": "Where the prompt sits among its kind and section; lower comes first"
          },
          "archived": {
            "type": "boolean",
            "description": "Whether the prompt is archived: no longer asked, its answers kept"
          },
          "archivedTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the prompt was archived"
          },
          "createdTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the prompt was created"
          },
          "lastUpdatedTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the prompt was last changed"
          }
        },
        "required": [
          "id",
          "kind",
          "section",
          "label",
          "position",
          "archived",
          "createdTime"
        ]
      },
      "ListReviewPromptsResponse": {
        "type": "object",
        "properties": {
          "reviewPrompts": {
            "description": "The caller's prompts: daily before weekly, Reflect before Plan, each in its order.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewPrompt"
            }
          }
        },
        "required": [
          "reviewPrompts"
        ]
      },
      "BaseReviewPrompt": {
        "type": "object",
        "properties": {
          "kind": {
            "description": "The kind of review that asks the prompt",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewKind"
              }
            ]
          },
          "section": {
            "description": "The step the prompt is asked in",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPromptSection"
              }
            ]
          },
          "label": {
            "type": "string",
            "description": "The question, 1 to 120 characters"
          },
          "placeholder": {
            "type": "string",
            "description": "A hint shown in the empty answer box, at most 200 characters"
          }
        },
        "required": [
          "kind",
          "section",
          "label"
        ]
      },
      "CreateReviewPromptRequest": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The prompt to add at the end of its kind and section.",
            "allOf": [
              {
                "$ref": "#/components/schemas/BaseReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
      "CreateReviewPromptResponse": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The prompt as created.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
      "ReorderReviewPromptsRequest": {
        "type": "object",
        "properties": {
          "kind": {
            "description": "The kind of review whose prompts are reordered",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewKind"
              }
            ]
          },
          "section": {
            "description": "The step whose prompts are reordered",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPromptSection"
              }
            ]
          },
          "promptIds": {
            "description": "Every one of the caller's prompt IDs of that kind and section, archived ones included, in the order they should appear.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "kind",
          "section",
          "promptIds"
        ]
      },
      "ReorderReviewPromptsResponse": {
        "type": "object",
        "properties": {
          "reviewPrompts": {
            "description": "All of the caller's prompts, in their new order.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewPrompt"
            }
          }
        },
        "required": [
          "reviewPrompts"
        ]
      },
      "DescribeReviewPromptResponse": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The prompt.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
      "PartialReviewPrompt": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "The question, 1 to 120 characters"
          },
          "placeholder": {
            "type": "string",
            "description": "A hint shown in the empty answer box, at most 200 characters"
          },
          "archived": {
            "type": "boolean",
            "description": "Whether the prompt is archived: no longer asked, its answers kept"
          }
        }
      },
      "UpdateReviewPromptRequest": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The changes to the prompt. A placeholder of null removes it; archived true archives it and false brings it back.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PartialReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
      "UpdateReviewPromptResponse": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The prompt with the changes applied.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
      "ReviewAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the answer"
          },
          "promptId": {
            "type": "string",
            "description": "The prompt the answer is to"
          },
          "body": {
            "type": "string",
            "description": "What was written, at most 10,000 characters"
          },
          "editedLater": {
            "type": "boolean",
            "description": "Whether the answer was written or last changed after its review was completed"
          },
          "createdTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the answer was first written"
          },
          "lastUpdatedTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the answer was last changed"
          }
        },
        "required": [
          "id",
          "promptId",
          "body",
          "editedLater",
          "createdTime"
        ]
      },
...
      "PartialReviewAnswer": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "description": "What was written, at most 10,000 characters"
          }
        }
      },
      "UpdateReviewAnswerRequest": {
        "type": "object",
        "properties": {
          "answer": {
            "description": "The answer as it should now read. A body that is empty or only whitespace removes the answer.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PartialReviewAnswer"
              }
            ]
          }
        },
        "required": [
          "answer"
        ]
      },
      "UpdateReviewAnswerResponse": {
        "type": "object",
        "properties": {
          "answer": {
            "description": "The answer as saved.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewAnswer"
              }
            ]
          }
        },
        "required": [
          "answer"
        ]
      }
    }
  }
}